Fixed bug in projection.py.Projection.getRaRanges()
The approximate center of the image was being calculated before the check
whether the image straddled the zero ra line. As a result, if the image
did straddle 0 degrees RA, the midpoint of the image was computed to lie
near 180 degrees. By delaying the computation of raMid until after that
check is done, the centre of the image is now computed correctly.